move signature, type signature, and volatility to split module (#1763)
